db_script/requirements.txt
elasticsearch
mysql-connector-python
requests
